Implemented Linux TIDs

+++ wyt/include/wyt.h
@@ -57,9 +57,10 @@ typedef signed long long wyt_duration_t;
 typedef void* wyt_thread_t;
 
 /**
- * Integer capable of holding a unique Thread Identifier.
+ * Integer capable of holding a Thread Identifier.
+ * A Thread ID is guaranteed to be unique at least as long as the thread is still running.
  */
-typedef unsigned long long wyt_tid_t;
+typedef unsigned int wyt_tid_t;
 
 // ================================================================================================================================
 //  API Functions
